requirements.txt
Django
django-storages
boto3
Pillow
gunicorn
